music_sanitize_names.sh: handle renaming conflicts (conservatively -- log conflicts).

+++ music_sanitize_names.sh
@@ -3,6 +3,11 @@ shopt -s extglob
 while read file; do
 	file_new="$(sed -re "s|\.*(\.[^\. ]+)?$|\1|" <<< "$file")"
 	if [[ "$file_new" != "$file" ]]; then
-		mv -v "$file" "$file_new"
+#		echo "'$file' -> '$file_new'"
+		if [[ -e "$file_new" ]]; then
+			echo "CONFLICT: '$file' -> already existing '$file_new'"
+		else
+			mv -v "$file" "$file_new"
+		fi
 	fi
 done < <(find . -mindepth 1 -depth)
